Update card photographer

+++ app/src/main/java/com/capstone/fotokui/ui/formphotographer/FormPhotographerViewModel.kt
@@ -53,7 +53,7 @@ class FormPhotographerViewModel @Inject constructor(
                 email = email,
                 experience = experience,
                 yearOrMonthExperience = yearOrMonthExperience,
-                price = price,
+                price = price.toFloat(),
                 promo = promo,
                 phone = phone,
                 description = description,
